+++ acvm/src/pwg/logic.rs
@@ -1,9 +1,9 @@
-use acir::circuit::gate::{AndGate, XorGate};
+use acir::circuit::gate::GadgetCall;
 use acir::native_types::Witness;
 use acir_field::FieldElement;
 use std::collections::BTreeMap;
 
-pub struct LogicSolver {}
+pub struct LogicSolver;
 
 impl LogicSolver {
     /// Derives the rest of the witness based on the initial low level variables
@@ -35,28 +35,33 @@ impl LogicSolver {
 
     pub fn solve_and_gate(
         initial_witness: &mut BTreeMap<Witness, FieldElement>,
-        gate: &AndGate,
+        gate: &GadgetCall,
     ) -> bool {
-        LogicSolver::solve_logic_gate(
-            initial_witness,
-            &gate.a,
-            &gate.b,
-            gate.result,
-            gate.num_bits,
-            false,
-        )
+        let (a, b, result, num_bits) = extract_input_output(gate);
+        LogicSolver::solve_logic_gate(initial_witness, &a, &b, result, num_bits, false)
     }
     pub fn solve_xor_gate(
         initial_witness: &mut BTreeMap<Witness, FieldElement>,
-        gate: &XorGate,
+        gate: &GadgetCall,
     ) -> bool {
-        LogicSolver::solve_logic_gate(
-            initial_witness,
-            &gate.a,
-            &gate.b,
-            gate.result,
-            gate.num_bits,
-            true,
-        )
+        let (a, b, result, num_bits) = extract_input_output(gate);
+        LogicSolver::solve_logic_gate(initial_witness, &a, &b, result, num_bits, true)
     }
+}
+// TODO: Is there somewhere else that we can put this?
+// TODO: extraction methods are needed for some opcodes like logic and range
+pub(crate) fn extract_input_output(gc: &GadgetCall) -> (Witness, Witness, Witness, u32) {
+    let a = &gc.inputs[0];
+    let b = &gc.inputs[1];
+    let result = &gc.outputs[0];
+
+    // The num_bits variable should be the same for all witnesses
+    assert_eq!(
+        a.num_bits, b.num_bits,
+        "number of bits specified for each input must be the same"
+    );
+
+    let num_bits = a.num_bits;
+
+    (a.witness, b.witness, *result, num_bits)
 }